feat: refactor to a single binary output
BREAKING CHANGE:
