+++ Cargo.toml
@@ -6,7 +6,11 @@
 [package]
 name = "aaprop"
 authors = ["Ali Sajid Imami <395482+AliSajid@users.noreply.github.com>"]
-categories = ["command-line-utilities", "science", "web-programming::http-server"]
+categories = [
+    "command-line-utilities",
+    "science",
+    "web-programming::http-server",
+]
 version = "1.1.1-next.2"
 description = "A server application to return the amino acid properties of a given amino acid."
 documentation = "https://docs.rs/crate/aaprop"
@@ -22,24 +26,25 @@ crate-type = ["dylib", "lib", "staticlib"]
 path = "src/aaprop_lib/lib.rs"
 
 [[bin]]
-name = "aaprop-server"
-path = "src/aaprop_server/main.rs"
-
-[[bin]]
 name = "aaprop"
-path = "src/aaprop_shuttle/main.rs"
+path = "src/aaprop/main.rs"
+
+[features]
+shuttle = ["dep:shuttle-axum", "dep:shuttle-runtime"]
+standalone = ["dep:clap", "dep:tokio", "dep:tracing-subscriber"]
+default = ["shuttle"]
 
 [dependencies]
 axum = "0.7.5"
-clap = { version = "4.5.4", features = ["derive", "env"] }
-serde = { version = "1.0.197", features = ["derive", "rc"] }
+cfg_block = "0.2.0"
+clap = { version = "4.5.4", features = ["derive", "env"], optional = true }
+serde = { version = "1.0.197", features = ["derive"] }
 serde_json = "1.0.115"
-shuttle-axum = "0.42.0"
-shuttle-runtime = "0.42.0"
-strum = { version = "0.26.2", features = ["derive", "strum_macros"] }
-strum_macros = "0.26.2"
-tokio = "1.37.0"
-tracing = { version = "0.1.38", features = ["async-await"] }
+shuttle-axum = { version = "0.42.0", optional = true }
+shuttle-runtime = { version = "0.42.0", optional = true }
+tokio = { version = "1.37.0", optional = true, features = ["full"] }
+tracing = "0.1.38"
+tracing-subscriber = {version = "0.3.18", optional = true}
 
 [dev-dependencies]
 rstest = "0.18.2"
@@ -48,18 +53,18 @@ rstest = "0.18.2"
 unsafe_code = "forbid" # Ensure that the crate does not use unsafe code
 
 [lints.rustdoc]
-broken_intra_doc_links = "deny" # Ensure that all intra-doc links point to valid items
-missing_docs = "forbid" # Ensure that all public items have documentation
-invalid_rust_codeblocks = "deny" # Ensure that all rust code blocks are valid
+broken_intra_doc_links = "deny"   # Ensure that all intra-doc links point to valid items
+missing_docs = "forbid"           # Ensure that all public items have documentation
+invalid_rust_codeblocks = "deny"  # Ensure that all rust code blocks are valid
 redundant_explicit_links = "deny" # Ensure that all explicit links are necessary
 
 [profile.dev]
 opt-level = 1 # Do not optimize
-debug = true # Generate debug symbols
+debug = true  # Generate debug symbols
 
 [profile.release]
-opt-level = 'z' # Optimize for size
-lto = true # Enable link-time optimization
+opt-level = 'z'   # Optimize for size
+lto = true        # Enable link-time optimization
 codegen-units = 1 # Compile the whole crate at once
-panic = "abort" # Abort on panic
+panic = "abort"   # Abort on panic
 strip = "symbols" # Strip debug symbols
