+++ Dockerfile
@@ -14,6 +14,9 @@
 # Use the 1.75 version of the Rust image since it's the MSRV (Minimum Supported Rust Version) for the aaprop project
 FROM rust:1.75 as builder
 
+# Install the required dependencies
+RUN apt-get update && apt-get install -y tree
+
 # Set the working directory in the builder image to /usr/src
 WORKDIR /usr/src
 
@@ -25,6 +28,7 @@ WORKDIR /usr/src/aaprop
 
 # Create the appropriate directory structure for the first build
 RUN mkdir -p src/aaprop_lib && mv -v src/lib.rs src/aaprop_lib/lib.rs
+RUN mkdir -p src/aaprop && touch src/aaprop/main.rs && echo "fn main() {}" > src/aaprop/main.rs
 
 # Copy the Cargo.toml and Cargo.lock files to the aaprop directory
 COPY Cargo.toml Cargo.lock ./
@@ -32,31 +36,31 @@ COPY Cargo.toml Cargo.lock ./
 # Build the Rust project
 # This step is done separately to take advantage of Docker's layer caching
 # Any changes in the source code will not invalidate the cached dependencies
-RUN cargo build --lib --release
+RUN cargo build --features standalone --no-default-features --release
 
 # Remove the auto-generated main.rs file
 # This file will be replaced with the actual source code
-RUN rm -rfv src/*
+# RUN rm -rfv src/*
 
 # Remove the auto-generated binary and dependencies
 # These will be replaced with the actual binary and dependencies
-RUN rm target/release/deps/aaprop*
+RUN rm -rfv target/release/deps/aaprop*
 
 # Add the actual source code to the src directory
 ADD src src
 
 # Build the Rust project with the actual source code
-RUN cargo build --bin aaprop-server --release --locked
+RUN cargo build --features standalone --no-default-features --release --locked
 
 # Use the official distroless image as the base image
 FROM gcr.io/distroless/cc-debian12
 
 # Copy the binary from the builder image to the base image
-COPY --from=builder /usr/src/aaprop/target/release/aaprop-server /usr/local/bin/aaprop-server
+COPY --from=builder /usr/src/aaprop/target/release/aaprop /usr/local/bin/aaprop
 
 # Change the user to a non-root user for security
 USER 1000
 
 # Set the binary as the entry point of the container
 # When the container starts, it will execute this binary
-ENTRYPOINT [ "/usr/local/bin/aaprop-server" ]
+ENTRYPOINT [ "/usr/local/bin/aaprop" ]
